refactor: move startAutoPosting to lib
